backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic-settings==2.7.1
PyJWT==2.9.0
bcrypt==4.2.0
pyodbc==5.1.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
